Changing to float

+++ configuration/pih/petl/jobs/mch/birth/target.sql
@@ -7,7 +7,7 @@ birth_number    INT,
 multiples       INT,
 birth_apgar     VARCHAR(10),
 birth_outcome   VARCHAR(100),
-birth_weight    DOUBLE,
+birth_weight    FLOAT,
 birth_neonatal_resuscitation    VARCHAR(10),
 birth_macerated_fetus           VARCHAR(10)
 );
